django-charts-api/requirements.txt
Django==5.0.3
Django==5.1.1
djangorestframework==3.15.2
